Fix problem reported by Mark Jordan in toRaw() output
We were using a test for the existence of data that stumbled upon ye
olde PHP interpretation of "string '0' is FALSE" which ended up
generating corrupted MARC binary output.

Add a unit test, using UUENCODE to avoid problems with binary file
encodings. You can decode it if you fear that it is not a valid MARC
file :)

Signed-off-by: Dan Scott <dan@coffeecode.net>