Fix incorrect insertField() behaviour
Our transition to SplDoublyLinkedList used the old insertNode()
implementation, which was overly complex and insufficiently tested.
As a result, an insertField() call could truncate a record.

Switch instead to the native SplDoublyLinkedList method for adding nodes
at the specified index, and a corresponding test. This simplifies the
code significantly and makes the library much more robust.

Thanks to Andreas Roussos for reporting this bug!

Signed-off-by: Dan Scott <dan@coffeecode.net>